:recycle: refactor markdown and mention handling functions in `__init__.py` and `public.py`
